@@ -1,21 +1,20 @@
-﻿namespace Atata.Bootstrap.Tests
+﻿namespace Atata.Bootstrap.Tests;
+
+using _ = DropdownPage;
+
+[Url("dropdown")]
+public class DropdownPage : Page<_>
 {
-    using _ = DropdownPage;
+    public RegularDropdown Regular { get; private set; }
 
-    [Url("dropdown")]
-    public class DropdownPage : Page<_>
+    public BSDropdown<_> Disabled { get; private set; }
+
+    public class RegularDropdown : BSDropdown<_>
     {
-        public RegularDropdown Regular { get; private set; }
+        public Link<_> Item1 { get; private set; }
 
-        public BSDropdown<_> Disabled { get; private set; }
+        public Link<_> Item2 { get; private set; }
 
-        public class RegularDropdown : BSDropdown<_>
-        {
-            public Link<_> Item1 { get; private set; }
-
-            public Link<_> Item2 { get; private set; }
-
-            public Link<_> Item3 { get; private set; }
-        }
+        public Link<_> Item3 { get; private set; }
     }
 }
