@@ -1,62 +1,56 @@
-﻿using System;
-using System.IO;
-using System.Linq;
-using System.Net.NetworkInformation;
-using System.Threading.Tasks;
+﻿using System.Net.NetworkInformation;
 using Atata.Cli;
 using Atata.WebDriverSetup;
-using NUnit.Framework;
 
-namespace Atata.Bootstrap.Tests
+namespace Atata.Bootstrap.Tests;
+
+[SetUpFixture]
+public class SetUpFixture
 {
-    [SetUpFixture]
-    public class SetUpFixture
+    private CliCommand _dotnetRunCommand;
+
+    [OneTimeSetUp]
+    public async Task GlobalSetUpAsync() =>
+        await Task.WhenAll(
+            DriverSetup.AutoSetUpAsync(BrowserNames.Chrome),
+            Task.Run(SetUpTestApp));
+
+    private static bool IsTestAppRunning() =>
+        IPGlobalProperties.GetIPGlobalProperties()
+            .GetActiveTcpListeners()
+            .Any(x => x.Port == UITestFixture.TestAppPort);
+
+    private void SetUpTestApp()
     {
-        private CliCommand _dotnetRunCommand;
+        if (!IsTestAppRunning())
+            StartTestApp();
+    }
 
-        [OneTimeSetUp]
-        public async Task GlobalSetUpAsync() =>
-            await Task.WhenAll(
-                DriverSetup.AutoSetUpAsync(BrowserNames.Chrome),
-                Task.Run(SetUpTestApp));
+    private void StartTestApp()
+    {
+        string testAppPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "..", "..", "..", "..", "Atata.Bootstrap.TestApp");
 
-        private static bool IsTestAppRunning() =>
-            IPGlobalProperties.GetIPGlobalProperties()
-                .GetActiveTcpListeners()
-                .Any(x => x.Port == UITestFixture.TestAppPort);
+        ProgramCli dotnetCli = new ProgramCli("dotnet", useCommandShell: true)
+            .WithWorkingDirectory(testAppPath);
 
-        private void SetUpTestApp()
+        _dotnetRunCommand = dotnetCli.Start("run");
+
+        var testAppWait = new SafeWait<SetUpFixture>(this)
         {
-            if (!IsTestAppRunning())
-                StartTestApp();
-        }
+            Timeout = TimeSpan.FromSeconds(40),
+            PollingInterval = TimeSpan.FromSeconds(0.2)
+        };
 
-        private void StartTestApp()
+        testAppWait.Until(x => IsTestAppRunning());
+    }
+
+    [OneTimeTearDown]
+    public void GlobalTearDown()
+    {
+        if (_dotnetRunCommand != null)
         {
-            string testAppPath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "..", "..", "..", "..", "Atata.Bootstrap.TestApp");
-
-            ProgramCli dotnetCli = new ProgramCli("dotnet", useCommandShell: true)
-                .WithWorkingDirectory(testAppPath);
-
-            _dotnetRunCommand = dotnetCli.Start("run");
-
-            var testAppWait = new SafeWait<SetUpFixture>(this)
-            {
-                Timeout = TimeSpan.FromSeconds(40),
-                PollingInterval = TimeSpan.FromSeconds(0.2)
-            };
-
-            testAppWait.Until(x => IsTestAppRunning());
-        }
-
-        [OneTimeTearDown]
-        public void GlobalTearDown()
-        {
-            if (_dotnetRunCommand != null)
-            {
-                _dotnetRunCommand.Kill(true);
-                _dotnetRunCommand.Dispose();
-            }
+            _dotnetRunCommand.Kill(true);
+            _dotnetRunCommand.Dispose();
         }
     }
 }
