@@ -1,39 +1,34 @@
-﻿using NUnit.Framework;
+﻿namespace Atata.Bootstrap.Tests;
 
-namespace Atata.Bootstrap.Tests
+[TestFixture("v3")]
+[TestFixture("v4")]
+[TestFixture("v5")]
+[Parallelizable]
+public abstract class UITestFixture
 {
-    [TestFixture("v3")]
-    [TestFixture("v4")]
-    [TestFixture("v5")]
-    [Parallelizable]
-    public abstract class UITestFixture
-    {
-        public const int TestAppPort = 59372;
+    public const int TestAppPort = 59372;
 
-        private readonly string _bootstrapVersionString;
+    private readonly string _bootstrapVersionString;
 
-        protected UITestFixture(string bootstrapVersionString) =>
-            _bootstrapVersionString = bootstrapVersionString;
+    protected UITestFixture(string bootstrapVersionString) =>
+        _bootstrapVersionString = bootstrapVersionString;
 
-        public static string BaseUrl { get; } = $"http://localhost:{TestAppPort}/";
+    public static string BaseUrl { get; } = $"http://localhost:{TestAppPort}/";
 
-        [SetUp]
-        public virtual void SetUp()
-        {
-            AtataContext.Configure()
-                .UseChrome()
-                    .WithArguments("window-size=1200,800")
-                    .WithArguments("headless")
-                .UseBaseUrl(BaseUrl + _bootstrapVersionString)
-                .UseCulture("en-US")
-                .UseNUnitTestName()
-                .LogConsumers.AddNUnitTestContext()
-                .LogNUnitError()
-                .Build();
-        }
+    [SetUp]
+    public virtual void SetUp() =>
+        AtataContext.Configure()
+            .UseChrome()
+                .WithArguments("window-size=1200,800")
+                .WithArguments("headless")
+            .UseBaseUrl(BaseUrl + _bootstrapVersionString)
+            .UseCulture("en-US")
+            .UseNUnitTestName()
+            .LogConsumers.AddNUnitTestContext()
+            .LogNUnitError()
+            .Build();
 
-        [TearDown]
-        public virtual void TearDown() =>
-            AtataContext.Current?.CleanUp();
-    }
+    [TearDown]
+    public virtual void TearDown() =>
+        AtataContext.Current?.CleanUp();
 }
